feat: complete registration logic (login logic pending)

+++ BackEnd/src/Utils/DeleteOne.js
@@ -4,6 +4,11 @@ const DeleteOne = async (model, filter) => {
 
   try
   {
+    if(!model || !filter)
+    {
+      return false;
+    }
+    
     const result = await model.deleteOne(filter)
     log(result)
     return result;
